TF all the infra things for easier future management (#151)
This is just to remove the reliance on the `hcloud` cli. It's handy and
all, but state can drift.

Ideally we would have some sort of remote state instead of requiring
encrypting/decrypting the tfstate, but TBH, it works so :shrug:

Here's hoping we don't have any quantum computers breaking SHA in the
next however long I have this hcloud account for :sweat_smile:.